+++ src/tab/TabGenerator.ts
@@ -44,6 +44,7 @@ export class TabGenerator extends Generator {
             let templateFiles = [
                 "src/app/scripts/{tabName}Config.tsx",
                 "src/app/scripts/{tabName}Tab.tsx",
+                "src/app/scripts/{tabName}Remove.tsx",
                 "src/app/web/{tabName}Tab.html",
                 "src/app/web/{tabName}Remove.html",
                 "src/app/web/{tabName}Config.html",
@@ -73,7 +74,7 @@ export class TabGenerator extends Generator {
             this.fs.writeJSON(manifestPath, manifest);
 
             Yotilities.addAdditionalDeps([
-                ["msteams-ui-components-react", "^0.4.7"],
+                ["msteams-ui-components-react", "^0.5.0"],
                 ["react", "^16.1.0"],
                 ["@types/react", "16.0.38"],
                 ["react-dom", "^16.2.0"],